refactor to use id Long

+++ Maki/src/main/java/com/maki/web/repository/RepositoryInterface.java
@@ -7,10 +7,10 @@ import java.util.Collection;
 
 public interface RepositoryInterface<T> {
     public Collection<T> selectAll();
-    public T selectById(Integer id) throws EntityNotFoundException;
+    public T selectById(Long id) throws EntityNotFoundException;
     public T insert(T entity) throws EntityConstraintException;
     public void delete(T entity) throws EntityNotFoundException;
-    public void deleteByID(Integer id) throws EntityNotFoundException;
+    public void deleteByID(Long id) throws EntityNotFoundException;
     public T update(T entity) throws  EntityConstraintException, EntityNotFoundException;
     public T upsert(T entity) throws  EntityConstraintException;
 }
